rename user pro_img testing

+++ users/views.py
@@ -188,15 +188,18 @@ def edit_profile(request):
 
         if profileform.is_valid() and userform.is_valid():
             if request.FILES:
-                profileform.save()
 
-                pro_img_directory = MEDIA_ROOT +'/'
-                os.remove(pro_img_directory + img_name)
-                os.rename(pro_img_directory + request.user.userprofile.pro_img.name,
+                if userprofile.pro_img:
+                    profileform.save()
+
+                    pro_img_directory = MEDIA_ROOT + '/'
+
+                    os.remove(pro_img_directory + img_name)
+                    os.rename(pro_img_directory + request.user.userprofile.pro_img.name,
                           pro_img_directory + request.user.username + '.png')
 
-                userprofile.pro_img.name = request.user.username
-                userprofile.save()
+                    userprofile.pro_img.name = request.user.username + '.png'
+                    userprofile.save()
             profileform.save()
             userform.save()
             request.session['django_timezone'] = request.POST['timezone']
